5. Подсчет функции в диапазоне.[#222238] Исправление

+++ chapter_004/src/test/java/ru/job4j/lambda/FunctionalRangeTest.java
@@ -28,7 +28,7 @@ public class FunctionalRangeTest {
     @Test
     public void whenExponentialFunctionThenExponentialResults() {
         FunctionRange function = new FunctionRange();
-        List<Double> result = function.diapason(5, 8, x -> (int) Math.pow(2, x));
+        List<Double> result = function.diapason(5, 8, x -> Math.pow(2, x));
         List<Double> expected = Arrays.asList(32D, 64D, 128D);
         assertThat(result, is(expected));
     }
